requirements.txt
PyQt6==6.9.0
mysql-connector-python==9.3.0
python-dotenv==1.0.1
